Handle console messages and render process exit
