+++ src/main/views/webContentEvents.ts
@@ -1,13 +1,11 @@
 // Copyright (c) 2016-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import path from 'path';
-
 import type {WebContents, Event} from 'electron';
 import {BrowserWindow, shell} from 'electron';
 
 import Config from 'common/config';
-import {Logger, getLevel} from 'common/log';
+import {Logger} from 'common/log';
 import ServerManager from 'common/servers/serverManager';
 import {
     isAdminUrl,
@@ -33,19 +31,14 @@ import ViewManager from 'main/views/viewManager';
 import CallsWidgetWindow from 'main/windows/callsWidgetWindow';
 import MainWindow from 'main/windows/mainWindow';
 
+import {generateHandleConsoleMessage} from './webContentEventsCommon';
+
 import {protocols} from '../../../electron-builder.json';
 import allowProtocolDialog from '../allowProtocolDialog';
 import {composeUserAgent} from '../utils';
 
 type CustomLogin = {
     inProgress: boolean;
-}
-
-enum ConsoleMessageLevel {
-    Verbose,
-    Info,
-    Warning,
-    Error
 }
 
 const log = new Logger('WebContentsEventManager');
@@ -302,27 +295,6 @@ export class WebContentsEventManager {
         };
     };
 
-    private generateHandleConsoleMessage = (webContentsId: number) => (_: Event, level: number, message: string, line: number, sourceId: string) => {
-        const wcLog = this.log(webContentsId).withPrefix('renderer');
-        let logFn = wcLog.debug;
-        switch (level) {
-        case ConsoleMessageLevel.Error:
-            logFn = wcLog.error;
-            break;
-        case ConsoleMessageLevel.Warning:
-            logFn = wcLog.warn;
-            break;
-        }
-
-        // Only include line entries if we're debugging
-        const entries = [message];
-        if (['debug', 'silly'].includes(getLevel())) {
-            entries.push(`(${path.basename(sourceId)}:${line})`);
-        }
-
-        logFn(...entries);
-    };
-
     removeWebContentsListeners = (id: number) => {
         if (this.listeners[id]) {
             this.listeners[id]();
@@ -362,7 +334,7 @@ export class WebContentsEventManager {
         // previously allowed from generateNewWindowListener through PluginsPopUpsManager.handleNewWindow.
         contents.on('did-create-window', PluginsPopUpsManager.handleCreateWindow);
 
-        const consoleMessage = this.generateHandleConsoleMessage(contents.id);
+        const consoleMessage = generateHandleConsoleMessage(this.log(contents.id));
         contents.on('console-message', consoleMessage);
 
         addListeners?.(contents);
